# openapi.yaml
openapi: 3.1.0
info:
  title: Human Design API
  version: 4.0.0
paths:
  /calculate:
    get:
      summary: Calculate Hd
      operationId: calculate_hd_calculate_get
      security:
      - HTTPBearer: []
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
          description: Birth year
          default: 1968
          title: Year
          example: 1968
        description: Birth year
      - name: month
        in: query
        required: true
        schema:
          type: integer
          description: Birth month
          default: 2
          title: Month
          example: 2
        description: Birth month
      - name: day
        in: query
        required: true
        schema:
          type: integer
          description: Birth day
          default: 21
          title: Day
          example: 21
        description: Birth day
      - name: hour
        in: query
        required: true
        schema:
          type: integer
          description: Birth hour
          default: 11
          title: Hour
          example: 11
        description: Birth hour
      - name: minute
        in: query
        required: true
        schema:
          type: integer
          description: Birth minute
          default: 0
          title: Minute
          example: 0
        description: Birth minute
      - name: second
        in: query
        required: false
        schema:
          type: integer
          description: Birth second (optional, default 0)
          default: 0
          title: Second
          example: 0
        description: Birth second (optional, default 0)
      - name: place
        in: query
        required: true
        schema:
          type: string
          description: Birth place (city, country)
          default: Kirikkale, Turkey
          title: Place
          example: Kirikkale, Turkey
        description: Birth place (city, country)
      - name: gender
        in: query
        required: false
        schema:
          type: string
          description: Gender (optional)
          default: male
          title: Gender
        description: Gender (optional)
      - name: islive
        in: query
        required: false
        schema:
          type: boolean
          description: Whether the person is alive (true) or deceased (false)
          default: true
          title: Islive
        description: Whether the person is alive (true) or deceased (false)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /bodygraph:
    get:
      summary: Get Bodygraph Image
      operationId: get_bodygraph_image_bodygraph_get
      security:
      - HTTPBearer: []
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
          description: Birth year
          default: 1968
          title: Year
        description: Birth year
      - name: month
        in: query
        required: true
        schema:
          type: integer
          description: Birth month
          default: 2
          title: Month
        description: Birth month
      - name: day
        in: query
        required: true
        schema:
          type: integer
          description: Birth day
          default: 21
          title: Day
        description: Birth day
      - name: hour
        in: query
        required: true
        schema:
          type: integer
          description: Birth hour
          default: 11
          title: Hour
        description: Birth hour
      - name: minute
        in: query
        required: true
        schema:
          type: integer
          description: Birth minute
          default: 0
          title: Minute
        description: Birth minute
      - name: second
        in: query
        required: false
        schema:
          type: integer
          description: Birth second (optional, default 0)
          default: 0
          title: Second
        description: Birth second (optional, default 0)
      - name: place
        in: query
        required: true
        schema:
          type: string
          description: Birth place (city, country)
          default: Kirikkale, Turkey
          title: Place
        description: Birth place (city, country)
      - name: fmt
        in: query
        required: false
        schema:
          type: string
          description: 'Image format: png, svg, jpg/jpeg'
          default: png
          title: Fmt
        description: 'Image format: png, svg, jpg/jpeg'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v2/calculate:
    post:
      summary: Calculate Human Design (V2)
      description: "Flagship V2 endpoint: Verbose by default, supports partial output masking, semantic enrichment, and advanced mechanics."
      operationId: calculate_hd_v2_post
      tags:
        - v2
      security:
        - HTTPBearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculateRequestV2'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculateResponseV2'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    CalculateResponse:
      type: object
      properties:
        general:
          $ref: '#/components/schemas/GeneralOutput'
        channels:
          type: array
          items:
            $ref: '#/components/schemas/ChannelInfo'
        gates:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/GateInfo'
      title: CalculateResponse
    GeneralOutput:
      type: object
      properties:
        birth_date:
          type: string
          title: Birth Date
        create_date:
          type: string
          title: Create Date
        birth_place:
          type: string
          title: Birth Place
        age:
          type: integer
          title: Age
          description: Calculated age in years
        gender:
          type: string
          title: Gender
          default: male
        islive:
          type: boolean
          title: Islive
          description: Whether the person is alive (true) or deceased (false)
          default: true
        zodiac_sign:
          type: string
          title: Zodiac Sign
          description: Western astrological sign
        energy_type:
          type: string
          title: Energy Type
        inner_authority:
          type: string
          title: Inner Authority
        inc_cross:
          type: string
          title: Inc Cross
        profile:
          type: string
          title: Profile
        active_chakras:
          type: array
          items:
            type: string
          title: Active Chakras
        inactive_chakras:
          type: array
          items:
            type: string
          title: Inactive Chakras
        definition:
          type: string
          title: Definition
        variables:
          $ref: '#/components/schemas/Variables'

      required:
      - birth_date
      - create_date
      - energy_type
      - inner_authority
      - inc_cross
      - profile
      - active_chakras
      - inactive_chakras
      - definition
      - variables
      title: GeneralOutput
    VariableDetail:
      type: object
      properties:
        value:
          type: string
          title: Value
        name:
          type: string
          title: Name
        aspect:
          type: string
          title: Aspect
        def_type:
          type: string
          title: Definition Type
      required:
      - value
      - name
      - aspect
      - def_type
      title: VariableDetail
    Variables:
      type: object
      properties:
        top_right:
          $ref: '#/components/schemas/VariableDetail'
        bottom_right:
          $ref: '#/components/schemas/VariableDetail'
        top_left:
          $ref: '#/components/schemas/VariableDetail'
        bottom_left:
          $ref: '#/components/schemas/VariableDetail'
        short_code:
          type: string
          title: Short Code
          description: Standard shorthand for all four variables (e.g., 'PRL DRR')
      required:
      - top_right
      - bottom_right
      - top_left
      - bottom_left
      - short_code
      title: Variables
    GateInfo:
      type: object
      properties:
        gate:
          type: integer
        line:
          type: integer
        color:
          type: integer
        tone:
          type: integer
        base:
          type: integer
        is_active:
          type: boolean
      required:
      - gate
      - line
      - color
      - tone
      - base
      - is_active
      title: GateInfo
    ChannelInfo:
      type: object
      properties:
        name:
          type: string
        gates:
          type: array
          items:
            type: integer
        meaning:
          type: string
      required:
      - name
      - gates
      title: ChannelInfo
...
    CalculateRequestV2:
      type: object
      required:
        - year
        - month
        - day
        - hour
        - minute
        - place
      properties:
        year:
          type: integer
          default: 1968
        month:
          type: integer
          default: 2
        day:
          type: integer
          default: 21
        hour:
          type: integer
          default: 11
        minute:
          type: integer
          default: 0
        second:
          type: integer
          default: 0
        place:
          type: string
          default: "Kirikkale, Turkey"
        gender:
          type: string
          default: "male"
        islive:
          type: boolean
          default: true
        latitude:
          type: number
          format: float
          nullable: true
        longitude:
          type: number
          format: float
          nullable: true
        include:
          type: array
          items:
            type: string
          nullable: true
          description: "List of sections to include. Supports dot-notation for nested fields (e.g. 'gates.personality')."
          example: ["general", "gates.personality"]
        exclude:
          type: array
          items:
            type: string
          nullable: true
          description: "List of sections to exclude. Supports dot-notation for nested fields."
          example: ["channels"]
      title: CalculateRequestV2
    CalculateResponseV2:
      type: object
      properties:
        general:
          $ref: '#/components/schemas/GeneralSectionV2'
        centers:
          $ref: '#/components/schemas/CentersV2'
        channels:
          type: array
          items:
            type: object
          nullable: true
        variables:
          $ref: '#/components/schemas/VariablesV2'
        gates:
          $ref: '#/components/schemas/GatesV2'
        mechanics:
          type: object
          nullable: true
        advanced:
          $ref: '#/components/schemas/AdvancedSectionV2'
      title: CalculateResponseV2
    GeneralSectionV2:
      type: object
      properties:
        birth_date:
          type: string
        create_date:
          type: string
        birth_place:
          type: string
        age:
          type: integer
        gender:
          type: string
        islive:
          type: boolean
        zodiac_sign:
          type: string
        energy_type:
          type: string
        strategy:
          type: string
        signature:
          type: string
        not_self:
          type: string
        aura:
          type: string
        inner_authority:
          type: string
        inc_cross:
          type: string
        profile:
          type: string
        definition:
          type: string
      title: GeneralSectionV2
    CentersV2:
      type: object
      properties:
        defined:
          type: array
          items:
            type: string
        undefined:
          type: array
          items:
            type: string
      title: CentersV2
    VariablesV2:
      type: object
      properties:
        top_right:
          $ref: '#/components/schemas/VariableItemV2'
        bottom_right:
          $ref: '#/components/schemas/VariableItemV2'
        top_left:
          $ref: '#/components/schemas/VariableItemV2'
        bottom_left:
          $ref: '#/components/schemas/VariableItemV2'
        short_code:
          type: string
      title: VariablesV2
    VariableItemV2:
      type: object
      properties:
        value:
          type: string
        name:
          type: string
        aspect:
          type: string
        def_type:
          type: string
      title: VariableItemV2
    GatesV2:
      type: object
      properties:
        personality:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/GateV2'
        design:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/GateV2'
      title: GatesV2
    GateV2:
      type: object
      properties:
        gate:
          type: integer
        line:
          type: integer
        color:
          type: integer
        tone:
          type: integer
        base:
          type: integer
        lon:
          type: number
        gate_name:
          type: string
          nullable: true
        gate_summary:
          type: string
          nullable: true
        line_name:
          type: string
          nullable: true
        line_description:
          type: string
          nullable: true
        fixation:
          type: object
          nullable: true
      title: GateV2
    AdvancedSectionV2:
      type: object
      properties:
        dream_rave:
          $ref: '#/components/schemas/DreamRaveOutput'
        global_cycle:
          $ref: '#/components/schemas/GlobalCycleOutput'
      title: AdvancedSectionV2
    DreamRaveOutput:
      type: object
      properties:
        activated_centers:
          type: array
          items:
            type: string
        activated_gates:
          type: array
          items:
            type: integer
        status:
          type: string
      title: DreamRaveOutput
    GlobalCycleOutput:
      type: object
      properties:
        great_cycle:
          type: string
        cycle_cross:
          type: string
        gates:
          type: array
          items:
            type: integer
        description:
          type: string
      title: GlobalCycleOutput
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer