# src/humandesign/schemas/response_models.py
# SPDX-License-Identifier: AGPL-3.0-or-later OR LicenseRef-DevAIble-Commercial
#
# Human Design API
# Copyright (C) 2026 Dogan Turkuler <dogan.turkuler@gmail.com>
# https://devaible.com
#
# This file is part of Human Design API, available under dual license:
#   - AGPL-3.0 (open source): see LICENSE-AGPL
#   - Commercial License: see LICENSE-COMMERCIAL or contact dogan.turkuler@gmail.com

from pydantic import BaseModel, Field
from typing import List, Dict, Optional

class VariableDetail(BaseModel):
    value: str
    name: str
    aspect: str
    def_type: str

class Variables(BaseModel):
    top_right: VariableDetail
    bottom_right: VariableDetail
    top_left: VariableDetail
    bottom_left: VariableDetail
    short_code: str = Field(..., description="Standard shorthand for all four variables (e.g., 'PRL DRR')")

class VariableSynergyDetail(BaseModel):
    alignment: str = Field(..., description="High-level alignment type (e.g., 'Symmetrical Force', 'Polarized')")
    operational_insight: str = Field(..., description="Business/Tactical insight based on arrow combination")
    lifestyle_insight: str = Field(..., description="Relational/Daily life insight based on arrow combination")
    shorthand_synergy: str = Field(..., description="Shorthand code for the synergy (e.g., 'Strategic-Heavy')")

class EnvironmentalResonanceDetail(BaseModel):
    resonance_type: str = Field(..., description="e.g. 'Shared Frequency', 'Harmonic Pull', 'Individual Path'")
    gates: List[int] = Field(..., description="The gates triggering the resonance")
    operational_insight: str = Field(..., description="Business context for the nodal environment")
    lifestyle_insight: str = Field(..., description="Relational context for the nodal environment")

class GeneralOutput(BaseModel):
    birth_date: str
    create_date: str
    birth_place: Optional[str] = None
    age: Optional[int] = Field(None, description="Calculated age in years")
    gender: Optional[str] = "male"
    islive: Optional[bool] = Field(True, description="Whether the person is still alive (True) or deceased (False)")
    zodiac_sign: Optional[str] = Field(None, description="Western astrological sign")
    energy_type: str
    inner_authority: str
    inc_cross: str
    profile: str
    active_chakras: List[str]
    inactive_chakras: List[str]
    definition: str
    variables: Variables

class GateInfo(BaseModel):
    gate: int
    line: int
    color: int
    tone: int
    base: int
    is_active: bool

class ChannelInfo(BaseModel):
    name: str
    gates: List[int]
    meaning: Optional[str] = None

class CalculateResponse(BaseModel):
    general: GeneralOutput
    channels: List[ChannelInfo]
    gates: Dict[str, List[GateInfo]]

class PentaDetail(BaseModel):
    active_skills: List[str]
    penta_gaps: List[str]
    is_functional: bool
    total_defined_centers: Optional[int] = Field(None, description="Total number of centers defined by the collective group")

class SubLineDetail(BaseModel):
    gate: int
    line: int
    color: int
    tone: int
    base: int
    planet: str
    position: str = Field(..., description="High-precision planetary position (e.g., '12.450')")

class MetaDetail(BaseModel):
    engine: str
    ephemeris: str
    timestamp: str

class MaiaDetail(BaseModel):
    channel: List[int] = Field(..., description="The two gates forming the channel")
    meaning: List[str] = Field(..., description="The spiritual and mechanical names of the channel")
    type: str = Field(..., description="Maia classification: Electromagnetic, Compromise, Dominance, or Companionship")
    circuitry: str = Field(..., description="The Human Design circuitry group (Individual, Tribal, Collective)")
    planetary_trigger: str = Field(..., description="The planets activating the gates in this connection")
    activations: List[SubLineDetail] = Field([], description="High-fidelity activation data (Gate, Line, Color, Tone, Base)")

class SynergyDetail(BaseModel):
    thematic_label: str = Field(..., description="Professional Maian theme (e.g. '8-1 Have some fun')")
    bridge_active: bool = Field(..., description="True if this combination bridges a split definition")
    center_dynamics: Dict[str, str] = Field(..., description="Conditioning dynamic for each center (fixed, conditioning_p1_to_p2, etc.)")
    aura_dynamic: str = Field(..., description="Professional label for Type-to-Type energetic interaction")
    love_gate_highlights: List[int] = Field(..., description="List of active G-Center love gates in the connection")
    space_count: int = Field(..., description="Number of completely open centers in the composite (Relational Space)")
    circuitry_dominant: str = Field(..., description="The dominant circuitry group in the partnership")
    profile_resonance: str = Field(..., description="Resonance or Harmony level between the two profiles")
    node_resonance: str = Field(..., description="Resonance level between the South and North Nodes (Environmental Fit)")
    dominant_sub_circuit: str = Field(..., description="Maximum detail on the sub-circuitry (e.g., Knowing vs. Logical)")
    planetary_flavor_summary: str = Field(..., description="The dominant planetary quality of the attraction (e.g., Venus-Mars)")
    group_dynamic_summary: str = Field(..., description="Classification of the group structure (Dyad, Penta, Wa)")
    penta_details: Optional[PentaDetail] = Field(None, description="Detailed Penta dynamics for groups of 3-5")
    variable_synergy: Optional[VariableSynergyDetail] = Field(None, description="Detailed analysis of Variable (Arrow) compatibility")
    environmental_resonance_detail: Optional[EnvironmentalResonanceDetail] = Field(None, description="Detailed nodal environment resonance analysis")

class CombinationItem(BaseModel):
    id: str
    other_person: str
    new_chakra: List[str]
    chakra_count: int
    new_channels: List[List[int]]
    new_ch_meaning: List[str]
    duplicated_channels: List[List[int]] = Field([], description="Channels already possessed by both participants individually (stability)")
    duplicated_ch_meaning: List[str] = Field([], description="Meanings of the duplicated channels")
    connection_code: str = Field(..., description="The 9-0 etc. connection classification code")
    maia_details: List[MaiaDetail] = Field(..., description="Deep Maian connection types for each channel")
    synergy: SynergyDetail = Field(..., description="Professional synergy and conditioning dynamics")

class CompositePersonDetail(BaseModel):
    name: str
    place: str
    tz: str
    birth_date: str
    create_date: str
    energy_type: str
    strategy: Optional[str] = None
    signature: Optional[str] = None
    not_self: Optional[str] = None
    aura: Optional[str] = None
    inner_authority: str
    inc_cross: str
    profile: str
    defined_centers: List[str]
    undefined_centers: List[str]
    definition: str
    variables: Optional[Variables] = None
    lunar_context: Optional[str] = Field(None, description="Lunar phase at birth")
    activations: Optional[Dict[str, SubLineDetail]] = Field(None, description="Full planetary activation matrix")
    channels: List[Dict]



class HybridAnalysisResponse(BaseModel):
    meta: Optional[MetaDetail] = Field(None, description="Engine provenance and calculation metadata")
    participants: Optional[Dict[str, CompositePersonDetail]] = Field(None, description="Full details of all participants")
    penta_dynamics: Optional[PentaDetail] = Field(None, description="Technical summary of the group Entity")
    dyad_matrix: List[CombinationItem] = Field(..., description="List of pairwise relationship analyses")